+++ src/main/java/com/jnape/palatable/lambda/adt/hlist/Tuple8.java
@@ -1,8 +1,10 @@
 package com.jnape.palatable.lambda.adt.hlist;
 
+import com.jnape.palatable.lambda.adt.Maybe;
 import com.jnape.palatable.lambda.adt.hlist.HList.HCons;
 import com.jnape.palatable.lambda.adt.product.Product8;
 import com.jnape.palatable.lambda.functions.Fn1;
+import com.jnape.palatable.lambda.functions.builtin.fn2.Into;
 import com.jnape.palatable.lambda.functor.Applicative;
 import com.jnape.palatable.lambda.functor.Bifunctor;
 import com.jnape.palatable.lambda.functor.builtin.Lazy;
@@ -10,6 +12,7 @@ import com.jnape.palatable.lambda.monad.Monad;
 import com.jnape.palatable.lambda.traversable.Traversable;
 
 import static com.jnape.palatable.lambda.functions.builtin.fn1.Constantly.constantly;
+import static com.jnape.palatable.lambda.functions.builtin.fn1.Uncons.uncons;
 
 /**
  * An 8-element tuple product type, implemented as a specialized HList. Supports random access.
@@ -248,5 +251,17 @@ public class Tuple8<_1, _2, _3, _4, _5, _6, _7, _8> extends HCons<_1, Tuple7<_2,
      */
     public static <A> Tuple8<A, A, A, A, A, A, A, A> fill(A a) {
         return tuple(a, a, a, a, a, a, a, a);
+    }
+
+    /**
+     * Return {@link Maybe#just(Object) just} the first eight elements from the given {@link Iterable}, or
+     * {@link Maybe#nothing() nothing} if there are less than eight elements.
+     *
+     * @param as  the {@link Iterable}
+     * @param <A> the {@link Iterable} element type
+     * @return {@link Maybe} the first seven elements of the given {@link Iterable}
+     */
+    public static <A> Maybe<Tuple8<A, A, A, A, A, A, A, A>> fromIterable(Iterable<A> as) {
+        return uncons(as).flatMap(Into.into((head, tail) -> Tuple7.fromIterable(tail).fmap(t -> t.cons(head))));
     }
 }