+++ src/test/java/com/jnape/palatable/lambda/adt/hlist/Tuple6Test.java
@@ -13,7 +13,12 @@ import testsupport.traits.FunctorLaws;
 import testsupport.traits.MonadLaws;
 import testsupport.traits.TraversableLaws;
 
+import static com.jnape.palatable.lambda.adt.Maybe.just;
+import static com.jnape.palatable.lambda.adt.Maybe.nothing;
 import static com.jnape.palatable.lambda.adt.hlist.HList.tuple;
+import static com.jnape.palatable.lambda.functions.builtin.fn1.Repeat.repeat;
+import static java.util.Collections.emptyList;
+import static java.util.Collections.singletonList;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -105,5 +110,12 @@ public class Tuple6Test {
         Tuple6<String, Integer, Integer, Integer, Integer, Integer>               a = tuple("foo", 1, 2, 3, 4, 5);
         Fn1<Integer, Tuple6<String, Integer, Integer, Integer, Integer, Integer>> b = x -> tuple("bar", 2, 3, 4, 5, x + 1);
         assertEquals(tuple("foo", 1, 2, 3, 4, 6), a.flatMap(b));
+    }
+
+    @Test
+    public void fromIterable() {
+        assertEquals(nothing(), Tuple6.fromIterable(emptyList()));
+        assertEquals(nothing(), Tuple6.fromIterable(singletonList(1)));
+        assertEquals(just(tuple(1, 1, 1, 1, 1, 1)), Tuple6.fromIterable(repeat(1)));
     }
 }